Bump up minor version, remove cljx hooks, add aliases
We're preparing for another release that should be CLJ/CLJS compatible,
so let's bump up the version to 0.2.0.

Remove the CLJX hooks for the time being - I don't have a good feel for
how cljx and cljsbuild interact, so let's remove the hooks for now. We
can always put them back.

Add aliases for testing the lib both on clj and cljs with one command,
and for deploying to clojars.